Remove schemas from Documents, add TypedDocument
This commit removes schemas from document types, which were previously
called Any. Instead, Document types assume the most appropriate and
generic schema from the prelude schama shapes when serialized.

To make a typed Document that wraps a shape, a special kind of Document
is created using Document.ofStruct. This method will serialize and
deserialize the Document exactly as if it was the underlying shape
(though codecs still have the opprortunity to change how documents are
serialized if necessary).