Serialize documents as document w/ separate serde
Documents are now serialized as a document when serialized through a
ShapeSerializer. The contents and inner data model contained within
a document can be accessed using Document#serializeContents. This
allows ShapeSerializer to receive documents as just documents and
the schema of a document matches the data model type written to the
serializer. It also allows serializers to, if they care, access the
contents of the document. This change makes documents more principled
and paves the way towards implementing validation without lots of
special casing.