Use consumers for serde state management
This is a partial revert of the last commit. The goal there was to
consolidate the logic to serialize a structure, but it ended up making
struct serialization different from list and map, and it also
introduced the need to manage struct state to ensure they're closed.
By using only the consumer based API, there is a clear hook for
before, during, and after that removes the need for state management.
It also does not have any measurable performance impact to use the
consumer API. Another benefit is that now StructSerializer can
be a FunctionalInterface.