Add validation
Validator is used to validate shapes. Validator does not itself throw
exceptions; this is left to the caller to validate a shape and do
something with the validation errors.

Validator validates the following:

- required members
- a single non-null member is set in a union
- pattern trait on strings
- enum values (the trait and the shape)
- intEnum values
- range trait on numbers
- length trait on strings, blobs, maps, and lists
- map keys and values at any level of nesting
- sparse trait for lists and map values
- stops validating when recursion limit is reached (default to 100)
- stops validating when error limit is reached (default to 100)

Note: uniqueItems is validated during deserialization. This simplifies
validation so that an entire list and its values don't have to be
loaded into memory to test uniqueness.