Add SerializableStruct for structure/union
Rather than rely on a BiConsumer to serialize structure and union
members, this change introduces a subtype of SerializableShape
specifically for structures and unions. With the BiConsumer approach,
in order to serialize a nested structure member, we had to reach
into the package-private inner serializers of a structure so that we
could serialize the member name from a member schema and then serialize
just the members of the target structure. This was a kind of implicit
contract and wouldn't work if the structure was from another package.

Now a structure/union shape exposes its schema using `schema()`,
a default method for serializing the structure/union via the inherited
`serialize()` method, and `serializeMembers()` to just write members
to a given serializer.

Performance is basically the same as before, though occasionally
jit inlining doesn't kick in.

A bug related to nested structure serialization in the JSON codec
was also fixed in this commit.

TODO: follow up commit to update codegen and remove the deprecated
method from ShapeSerializer.