Remove schema from public API and clean up
Schema wasn't being used by anything other than the default impl of
SerializableStruct. Removing it for now to encourage using generated
classes rather than relying on schemas. We can add it back later if
we really need it, but for now it's safer to remove it.

This commit also removes the methods from ShapeSerializer that
accept a value without a schema (struct and document methods). This
makes it easier to implement the interface, and the only places the
methods that didn't provide a schema were being used was in tests.