Rely on ShapeSerializer for stream bindings
We previously relied on externalized support for serializing event/
data streams by extracting them from inputs and explicitly passing
them to call and clientCall. Now that this is handled in
ShapeSerializer, this is no longer necessary. We don't need to allow
them to be set in HttpBinding classes either and can just rely on
the shape serialization.