Make JSON parsers pluggable, default to Jackson
Despite json-iter's superior performance, it had some questionable
behavior when handling untrusted input. This adds Jackson as a default
implementation with json-iter as an available opt-in, and the ability
for additional runtime customization of JSON parsing. It would be neat,
for instance, to add a simd-json module for apps running on JDK 21.