Add support for getting document discriminators
Typed documents contain a protocol-specific discriminator. For
example, a JSON document provided by the JsonCodec uses "__type".
Protocols might return an absolute shape ID or a relative shape ID.
To abstract over these differences, a `discriminator()` method has
been added to Document type find the discriminator and parse it into
a valid shape ID. If you attempt to get the discriminator of the
wrong kind of document, or the document doesn't have a discriminator,
or there's something wrong with the shape ID syntax, then a
DiscriminatorException is thrown.

Attempting to get the discriminator of a document is an opt-in choice,
so rather than return null and still throw in some cases (like when
a shape ID has a syntax exception), the API never returns null and
throws when it can't return a shape ID.

To allow wiring up a default namespace for parsing discriminators from
relative shape IDs with the JSON codec, an optional "defaultNamespace"
property has been added to the rest-json protocol. Clients should
configure this when adding this protocol to work with relative
shape ID discriminators.

For now, the way this would work with a type registry is:

```
shape = document.asShape(registry.getBuilder(document.discriminator()));
```