Add size hints to documents and deser
Documents and ShapeDeserializers can now provide the size of a
container. This information can be used to create exact-sized
collections to deserialize values into, avoiding over or under
allocating for protocols that know the size of data to be read.
For documents, it's the size of a list/map or -1 if not a list/map.
For ShapeDeserializer, it's the size of a list/map about to be read,
or -1 if unknown or not a list/map.