Make JSON perf improvements
Disables string interning since it locks and isn't useful for us.

Make use of parser.getNextFieldName() to improve deserializing maps
and structures.

Use a switch when deserializing floats and doubles.

Use the fast double and big-number parsers.

Add a cache of SerializeString values to use when serializing structure
member names.

Remove the use of Map.computeIfAbsent from a few places
in the JSON codec package.

Add a codec method for deserializing from a ByteBuffer.

Add cache for SerializableString. The built-in interning really hurts
performance for maps, but we still want interning for structure and
union members.