Add AWS JSON protocols and error deserializer
This commit adds AwsJson protocol support. The restJson1 implementation
was moved into the same package as this protocol and moved under the
aws package. They're both just a single class with the same dependency
tree.

A generic error deserialization library was added that is used by
both awsJson and restJson, and could be used by likely most HTTP
based protocols, including protocols that use HTTP bindings.

This commit also updates structure and union error correction to
tell the presence tracker that they are set even when they aren't.
This causes required structures and unions to default to null, which
matches Smithy spec.