Move DataStream to common and use in http-api
To reduce the need to convert to and from a Flow.Publisher when the
data in an HTTP message is already a DataStream, this commit moves
DataStream and a few other classes from core into a common package
(more things should be moved here soon, like Context and Logging).

This has a significant performance benefit as it reduces copies and
lots of method call overhead to convert from a Publisher to a
DataStream.