Add faster TraitMap implementations
It looks like we don't need an enum-based trait map for prelude traits
with these custom TraitMap classes for common cases of 0-5 traits.
An enum based trait map are about 0.9 ns/op, and we're able to exceed
that with O(N) identity lookups of classes stored in local variables
up until around the 4th or 5th element, so the return on adding an
entirely different access pattern for only prelude traits doesn't
make as much sense.