Dont add docs to client impls, make docs interface
There's now an interface, DocumentedSection, that's used to trigger
the javadoc interceptor. The behavior of this interface can now be
configured with the ApplyDocumentation enum. This enum is used to
ensure that javadocs are not injected for implementations of client
classes, since those should just implicitly inherit the javadocs of
the client interface.