Unconditionally index into trait map
This change brings performance back to where it was before making the
sliding style trait map array. It's done by adding null padding to the
beginning and end of the trait array and clamping the adjusted index
so that out of bounds indices land on the padding and return null.