Add MemberLookup to reduce map lookups in serde
When serializing a struct or union with JSON and jsonName traits, the
JsonFieldMapper required a CHM lookup for the containing structure and
then a HashMap lookup for each member, over and over. This change
introduces a MemberLookup interface that Schema implements, and it's
used to lookup members by name. When deserializing structs without
JSON name, the JsonFieldMapper just uses the Schema directly, so
there's no perf hit in this PR. When deserializing with jsonName,
JsonFieldMapper first does a cache lookup in CHM, then gets the
container for the structure, and uses it for each member to
deserializing, turning the N cache lookups for the parent container
to 1 lookup.

When member lookups by char/byte are added for faster CBOR and JSON
parsing, we can add those canonicalizing and lookup methods to
MemberLookup too so that looking up a member using the jsonName trait
is just as fast as looking one up using member name (plus the single
CHM lookup to cache the computed table).