Add HttpHeaders abstraction
Consolidate the client server parts of the codebase to both rely on
an HttpHeaders abstraction that is _not_ the one from the JDK
HttpClient. The built-in HttpHeaders uses Optionals and always requires
a predicate to create headers. We also wanted a mutable variant for the
server side, but immutable for the client side.

I've tweaked the existing interface to look more like the built-in
HttpHeaders Java interface to reduce churn.