Add error deserialization to DynamicClient
DynamicClient can now deserialize errors dynamically with a schema and
return the document contents of the error. This required passing the
service shape around to each thing that can create a client.

This also includes a bugfix for AwsJson protocols so that it actually
passed the default service namespace through to the underyling
JSON codec.