Treat clientOptional as present
ClientOptional required members not being present won't block the
creation of a POJO, but will still fail validation. We achieve this by
just telling the tracker that the member is present in the builder
ctor.

The idempotency token injector now works with required tokens, and it
will now treat an empty string token value as not set and replace it.