Start to move away from ByteBuffer to Slice
ByteBuffer has tricky state in it's position and cursor functionality.
We don't use any of that in smithy-java, so we often forget to account
for moving or not moving the position. This commit introduces Slice,
which is basically a lighter ByteBuffer with no state. We'll add
more commits that migrate other abstractions to Slice.