Treat ClientTransport + MessageExchange as plugins
ClientTransport now implements ClientPlugin so that when it's applied,
it can configure the client to correctly behave for the selected
transport. MessageExchange also now is a ClientPlugin and is used to
configure the default behavior of a client for a given exchange
(e.g., HTTP should use the User-Agent plugin by default). The default
implementation of ClientTransport will automatically apply the
messageExchange()'s configureClient method to apply exchange-wide
configuration. When overriding configureClient on a ClientTransport,
you can choose to completely exclude the default modifications applied
by the MessageExchange or further augment any changes made by the
MessageExchange (taking care to call configureClient on the
MessageExchange manually if desired).

The ClientTransport configuraiton is applied when ClientConfig#build
is called. This is also the place where ClientTransport is resolved
via SPI, so it makes sense here. However, because a config can be
converted to a builder and rebuilt, we now need to keep a copy of the
original builder that was used to create the config. This prevents
configuration modifications of a transport from taking effect multiple
times, or from a config that changes to a different transport from
still retaining the modifications of the previous builder
configuration.