Move uniqueItems validation to Validator
Rather than validate unique items in list shape creation, we now
validate it only during validation. This allows clients and servers to
losslessly deserialize even lists with unique items and doesn't require
any kind of potential modality in deserialization between client and
server. It also means that POJOs don't throw framework errors, which
we're trying to decouple from the core libraries. Finally, this now
allows things like DynamicCLient and documents to validate uniqueItems.

To implement uniqueItems, we create a set of documents and emit an error
on each list index that is a duplicate of a previous index.