Add our own exceptions for ClientTransport
To avoid directly exposing the exceptions of the underlying transport of
a ClientTransport, which inadvertently allows for coupling of caller
logic to a specific transport, we now provide a set of transport exceptions
that extend from CallException.

ClienTransports are expected to only throw a subtype of CallException or
ClienTransportException, and to wrap other exceptions. ClientPipeline
has been updated to also perform this remapping in case a transport
either is implemented incorrectly or tries to work around this mapping
(which again allows for coupling to a specific transport).