Add service schema to clients
Clients now get access to the schema of the service, providing the shape
ID and traits of the service allowed for code generation. This can be
used to know what service is being called for telemetry, or implement things
like endpoint resolvers or custom auth handling based on traits applied to a
service.

Each generated ApiOperation now provides a getter called `service()` to
get the generated ApiService for the client. Resources get this method
too. The generated ApiService only provides the schema of the service.
It doesn't provide a full accounting of every operation or resource
within the service as there isn't a clear use for that, it adds
complexity due to bidirectional references, and we want to avoid eagerly
loading every shape in a service when the ApiService is loaded.
This commit also simplifies the generated ApiResource for this same
reason, removing the operation and sub-resource schemas.

DynamicClient will now automatically generate an ApiService based on the
Smithy model.