Revert to using RequestOverrideConfig to configure the dynamic client
ClientPlugin does not currently offer the level of flexibility I need to
dynamically configure the client on a per-request basis. Namely, it is
extremely difficult (if not impossible) to configure the protocol in
this manner. Protocols are explicitly tied to their transports, meaning
I don't have a good opportunity to swap one out while retaining the
other after the internal ClientCall object has been constructed.
RequestOverrideConfig is the only extension point in the client today
that will allow me to influence the ClientCall itself, and therefore it
presents the most consistent and reliable mechanism for overriding every
property that I need to control on each call.