Expose framework error classes to server dependencies
When compiling a project that uses the server-rpcv2-cbor package we get
an error that Exception classes declared in the framework-error
package cannot be found. server-rpcv2-cbor declares an api dependency
on server-api which is a common package for server related code. If
framework-errors are intended to be common server logic then server-api
should declare an api dependency on framework-errors so that consumers
will be able to access those classes (which are used in code generation).