Add a way to set CUSTOM_ENDPOINT and use with EP 2
Client builder now has new `endpoint` builder methods that are used
to more easily set a custom endpoint for the client. This also sets the
ClientContext.CUSTOM_ENDPOINT setting that the rules engine plugin
looks for two (1) know if it should override the currently set
endpoint provider of the client (2) what the SDK::Endpoint builtin
should return when called.

This commit also breaks up some of the context properties that were
all stored in CallContext and moves some that can be client-wide to
ClientContext.