Add support for @eventHeader and @eventPayload for RPC protocols
Adds support to encode and decode members with the @eventHeader and
@eventPayload traits. For the decoding part, it's necessary to decode
the union member independently of the union itself for which we need
its builder, but we currently have no means to get one. For this, the
schema was enhanced wiht a method that to return a supplier to
generate new instances of builder for that schema which is only set
for structures and unions.